tests/wpt/tests/css/css-ui/parsing/user-find-invalid.tentative.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS UI Level 4: parsing user-find with invalid values</title>
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/3460">
<meta name="assert" content="user-find supports only the grammar 'auto | none'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("user-find", "visible");
test_invalid_value("user-find", "never");
test_invalid_value("user-find", "10px");
test_invalid_value("user-find", "'none'");
</script>
</body>
</html>
